requirements.txt
fastapi==0.109.0
ultralytics==8.0.232
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.3
uvicorn==0.27.0
pillow==10.2.0